flash sale process

+++ FlashSale/src/main/java/com/ecommercesystemtemplate/flashsale/to/FlashSaleSkuRedisTo.java
@@ -30,7 +30,7 @@ public class FlashSaleSkuRedisTo {
     /**
      * flash limit
      */
-    private BigDecimal seckillLimit;
+    private Integer seckillLimit;
     /**
      * flash sort
      */
